+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 A Flutter widget that lays out its children in tight rows.
 
+## [1.0.2]
+
+### Added
+
+- More test cases for AStar
+
+### Fixed
+
+- A couple of issues with gallery sizing in AStar when the gallery is very narrow
+- Prevent AStar gallery from reporting infinite height
+
 ## [1.0.1]
 
 ### Added
@@ -57,6 +68,8 @@ A Flutter widget that lays out its children in tight rows.
 - CircleCI config
 
 [Unreleased]: https://bitbucket.org/mongoose13/fling-gallery/commits/
+[1.0.2]: https://bitbucket.org/mongoose13/fling-gallery/commits/tag/1.0.2
+[1.0.1]: https://bitbucket.org/mongoose13/fling-gallery/commits/tag/1.0.1
 [1.0.0]: https://bitbucket.org/mongoose13/fling-gallery/commits/tag/1.0.0
 [0.2.0]: https://bitbucket.org/mongoose13/fling-gallery/commits/tag/0.2.0
 [0.1.0]: https://bitbucket.org/mongoose13/fling-gallery/commits/tag/0.1.0